fix(image-edit): fix event name disappearing
The event title disappeared because the image caption field's name attribute
used the "name" value and was defined as a HelTextField element. The
HelTextField element updates the store on the blur event and because the
name of the caption field was "name", it collided with the event's title
which is also stored in the store under the "name" key. Thus, the image
caption field overrode the event title. Since the event title is an object
and the image caption just a string, the event title appeared as an empty
field in the UI.

There doesn't seem to be any reason to use HelTextField elements for the
image caption and alt text fields since validation isn't necessary. The
save button is already disabled if the fields contain invalid data. Also,
there's no need to store the image metadata in the store which the
HelTextField automatically does on the blur event.