fix(image-edit): fix image metadata field validations
Revert the fix done in commit 482f47e855b3c590a7b042bbbb71aaf4d5754585.
In that fix, HelTextFields were replaced by normal TextFields because we
didn't want to use the store with the image metadata fields.

However, HelTextFields contain functionality for showing how many
characters a user can still type in a field (an indicator for the
validity of the input data). By replacing HelTextFields with TextFields,
this functionality was also removed.

To fix this, we start using HelTextFields again but change their inner
logic so that the user can decide whether to store any changes to the
store or not. In the long run, it might be best to move the store logic
from these kind of low level components to parent components so that
the user can decide what to do with the data.