Feature/22 event location map (#96)
* Added EventMap component that displays the event location on the editor page via the formfields component. The map component visibility is toggled by a button, a tooltip is displayed on the button if a location has not been selected.
Updated FormFields/index with the new EventMap component and a button that toggles its visibility.
Updated HelSelect with a new prop 'optionalWrapperAttributes' that are used to add attributes to the div that wraps the Select, added aria-labelledby to the Select component.
Major rework of HelSelect functions that are used when selecting location, search results are now displayed in the currently selected language, if the selected location is missing ie. a swedish name then the finnish one is used by default.
Added new language specific messages that are used by the EventMap component/toggle button.
Added two json files to src/assets/default, these are used when city_theme is not in use, urls has the url to the helsinki tileset and localization contains the default map center/position for helsinki.

* Added aria-label to toggle button that is used if no event location is selected, added className to toggle button which adds disabled state styling to the button if no event location is selected.
Removed tooltip from the button as it is no longer needed with the new disabled state styling added to the button.