Added changes to make the editor page display input texts according to the selected language. (#97)
Changes include:

EventDetails/index.js: Changed OptionGroup function to display event checkbox values according to selected language.

FormFields/index.js: Added availableLocales array[string] to component state that contains the locales that can be specified for data retrieval. Added currentLocale variable that is passed to location/keyword select that checks if current intl.locale is included in availableLocales.

HelKeywordSelector.js: Category texts change according to the selected language, currentLocale is passed through to other components.

HelSelect.js: Keyword search results are displayed according to the selected language, reworked the location search so that it is similar to the new keyword search.

ImagePicker/index.js: Changed order of exports, this seems to have fixed the texts that didnt change with language.

SelectedKeywords.js: Text changes according to the selected language.

sv.json: Added a bunch of translations that were previously missing. Some of the translations are quite rough, change them when a better translation is provided.

apiDataMapping.js: Added an additional key to the return object of the mapKeywordSetToForm function.

formDataMapping.js Modified obj.keywords so that if you click 'edit event' on an event page, the following editor page has all of the language specific keywords